customHistoryMenu implemented and bug fixes
Implement customHistoryMenu to allow user to enter custom histories.
History class now has writeOutHistory that writes back its operations as a string.

Bug Fixes:
Addressed bug fix in HistoryCreator: should be greater than, not greater than or equal to maxIDs.
Transaction should not be created if it does not have a commit or abort.
Addressed ACAAnalyzer and STAnalyzer to output that the History is Strict when no transaction reads or writes to a data item that is not used by any other transaction and ACA when no transaction reads a data item not used by another transaction.